Revert "loading default data for stock price and who owns shares"
This reverts commit 99167b74e206cf55970ed854f7264de4e37bbd49.

+++ src/main/java/ie/atu/cicd_project/StockValueService.java
@@ -10,7 +10,7 @@ public class StockValueService {
 
     public void portfolioResult(int stockShares, String stockName){
         StockValue stockValue = stockValueRepository.findByStockName(stockName);
-        double ShareVal = stockValue.getStockShareValue();
+        int ShareVal = stockValue.getStockShareValue();
         System.out.println((float) (ShareVal*stockShares));
     }
 }
